% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze.contrast.R
\name{analyze.contrast}
\alias{analyze.contrast}
\title{Takes in a dds DESeq2 object and extracts results for a specific contrast.}
\usage{
analyze.contrast(dds, control_cond, experiment_cond, gene_features,
  contrast_name = "", celltype = "", species = "unknown_species",
  date = date())
}
\value{
d
}
\description{
This function was build for the repetitive contrast done in the Chuck Harrell Brain Metastasis project.
It takes in severla parameters, extracts results for a specified contrast, saves the data, plots and saves a MA Plot, and plots the expression values of the top DE gene.
Note: this assumes the dds condition column is named "condition" and the gene_features gene symbol column is named "Gene.Symbol".
}
\examples{
data <- analyze.contrast()
}
\author{
Amy L. Olex \email{alolex@vcu.edu}
}
\keyword{d}
