% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc.percentSpeciesExpression.R
\name{calc.percentSpeciesExpression}
\alias{calc.percentSpeciesExpression}
\title{Calculates the percent of gene expression for each listed species.}
\usage{
calc.percentSpeciesExpression(data, species_df)
}
\arguments{
\item{data}{A dataframe or matrix containing the data to use for calculating the percent of species expression.}

\item{species_df}{A dataframe containing each species name as a column name, and the string to use to identify each species uniquely in the gene names(row names of data matrix) is in row 1.}
}
\value{
dataframe with one row per sample/condition and N columns with the percent of expression for the provided species.
}
\description{
This function calculates the percent of gene expression for each listed species base on the total read counts and the counts of reads belonging to each species.
This function requires that the gene names for each species be uniquily labeled.  For example, ENSG is human and ENSMUSG is mouse for Entre IDs.
}
\examples{
species_df <- data.frame("Human"="ENSG", "Mouse"="ENSMUSG")
percents <- calc.percentSpeciesExpression(data, species_df)
}
\author{
Amy L. Olex \email{alolex@vcu.edu}
}
\keyword{expression,}
\keyword{gene,}
\keyword{percent}
\keyword{species}
