% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center.palette.R
\name{center.palette}
\alias{center.palette}
\title{Calculates a color palette where zero values are centered at white.}
\usage{
center.palette(data, palette_length = 100, color1 = "blue",
  color2 = "red")
}
\arguments{
\item{data}{A matrix of integers.}

\item{palette_length}{The size of the color palette. Default = 100}

\item{color1}{The color for the low bound. Default = "blue"}

\item{color2}{The upper bound color.  Default = "red"}
}
\value{
A vector containing the new color palette.
}
\description{
Calculates a color palette where zero values are centered at white.
}
\examples{
my_palette <- center.palette(data, palette_length = 200, color1 = "black", color2 = "pink")
}
\author{
Amy L. Olex \email{alolex@vcu.edu}
}
\keyword{color,}
\keyword{heatmap,}
\keyword{palette}
