% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.to.colors.R
\name{convert.to.colors}
\alias{convert.to.colors}
\title{Converts a data frame of factors into a list of numerical vectors representing colors.}
\usage{
convert.to.colors(annot.df)
}
\arguments{
\item{annot.df}{A data frame where each column is a factor.}
}
\value{
A list of vectors with numbers representing colors, one list item per column in the input data frame.
}
\description{
This function converts the input data frame to a list of numerical vectors that can be input into the aheatmap() function for column and row annotations.
}
\examples{
my_colors <- convert.to.colors(annotation.df)
}
\author{
Amy L. Olex \email{alolex@vcu.edu}
}
\keyword{annotation}
\keyword{color,}
