% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deep.heatmap.R
\name{deep.heatmap}
\alias{deep.heatmap}
\title{Creates an annotated heatmap from DeepTools multiBamSummary tab delimited output.}
\usage{
deep.heatmap(deep_file, annot_file, outfile_name = "image.jpg",
  plot_title = "")
}
\arguments{
\item{deep_file}{The path and name of the file containing the DeepTools tab delimited results from the multiBamSummary method.}

\item{annot_file}{The path and file name of the tab delimited text file containing the sample annotations to use for adding colored annotation bars on the heatmap.}
}
\value{
A data frame of bin counts from the multiBamSummary output, and a data frame with annotations.
}
\description{
This function dreated a heatmap from the DeepTools multiBamSummary method, and adds color coded annotation bars to the heatmap to annotation samples.
The file is formatted with the following fields: #'chr'	'start'	'end'	'sample1_name.bam' 'sample2_name.bam' ... 'sampleN_name.bam'
}
\examples{
data <- deep.heatmap(deep_file="counts.tab", annot_file="annots.txt", image_name="Plot Title")
}
\author{
Amy L. Olex \email{alolex@vcu.edu}
}
\keyword{deeptools,}
\keyword{heatmap}
\keyword{import,}
