% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.removeZeroSumRows.R
\name{filter.removeZeroSumRows}
\alias{filter.removeZeroSumRows}
\title{Removes all rows of a matrix or dataframe where the row sum is zero.}
\usage{
filter.removeZeroSumRows(data)
}
\arguments{
\item{data}{A dataframe or matrix containing the data to reduce.}
}
\value{
A matrix of raw read counts without rows that summed to zero
}
\description{
This function reduces an RNAseq data set by removing all rows that sum to zero.  These rows represent genes that were not expressed in any sample/condition.
}
\examples{
data <- filter.removeZeroSumRows(data)
}
\author{
Amy L. Olex \email{alolex@vcu.edu}
}
\keyword{expression,}
\keyword{filter}
\keyword{gene,}
