% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.subreadFormat.R
\name{load.subreadFormat}
\alias{load.subreadFormat}
\title{Imports raw read counts from SudRead's featureCounts() formatted output.}
\usage{
load.subreadFormat(file_list, annotations = NA, annot_id = 1)
}
\arguments{
\item{file_list}{A list of file names for concatenate together into a single read count matrix.}

\item{annotations}{Optional. Name of a file with the gene feature annotations in tab delim format. First column is the Gene Official Symbol and second column is the Gene ID present in the read count files.}

\item{annot_id}{Optional. The column number that contains the feature IDs that match the read count files.  The other column must be the wanted gene identifiers like official gene symbol. Default is 1.}
}
\value{
A matrix of raw read counts.
}
\description{
This function iterates through one or more text files that contain read counts obtained from SubRead's featureCounts() method.
Multiple files MUST have the same feature set with different sample columns.
This method extracts the count data and concatenates the matricies into one matrix.  It also maps the gene symbols, if provided, to the gene ID.
This method ASSUMES the first 5 columns of the data frame (not including the row names) are the annotation data for Chr, Start, End, Strand, and Length.It only keeps the information from the first file and deletes the first 5 columns of all the other files before concatenating.
}
\examples{
data <- load.subreadFormat(file_list = c("file1","file2","file3"), annotations="file", annot_id=2)
}
\author{
Amy L. Olex \email{alolex@vcu.edu}
}
\keyword{expression,}
\keyword{gene,}
\keyword{import}
