% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTPM.R
\name{calc.tpm}
\alias{calc.tpm}
\title{Calculates the TPM values for RNA-seq data.}
\usage{
calc.tpm(data)
}
\arguments{
\item{data}{A matrix of read count and gene length data with genes as rows and samples as columns.  The very first column of the matrix MUST be the gene lengths, and MUST be named "Lengths".}
}
\value{
A matrix of TPM expression values.
}
\description{
This function uses the method described at () to calculate the TPM (Transcripts Per Million) values from raw RNA-seq read counts.
}
\examples{
tpm <- calc.tpm(data, features)
}
\author{
Amy L. Olex \email{alolex@vcu.edu}
}
\keyword{expression}
\keyword{gene,}
\keyword{tpm,}
